refactor: move all test to one folder
